+++ src/lib/hooks/index.ts
@@ -6,6 +6,6 @@ export * from "./useRedo";
 export * from "./useSingleMessageProps";
 export * from "./useToast";
 export * from "./useUserKey";
-export * from "./useOpenTxTab";
+export * from "./useOpenTab";
 export * from "./useIsCurrentPage";
 export * from "./useGetMaxLengthError";
