--- src/lib/hooks/useOpenAssetTab.ts
@@ -1,16 +1,0 @@
-import { useCallback } from "react";
-
-import { useBaseApiRoute } from "lib/app-provider";
-
-import { openNewTab } from "./useOpenTxTab";
-
-export const useOpenAssetTab = () => {
-  const balancesApiRoute = useBaseApiRoute("balances");
-
-  return useCallback(
-    (walletAddr: string) => {
-      openNewTab(`${balancesApiRoute}/${walletAddr}`);
-    },
-    [balancesApiRoute]
-  );
-};
