+++ src/lib/utils/window.ts
@@ -1,0 +1,2 @@
+export const openNewTab = (url: string) =>
+  window.open(url, "_blank", "noopener,noreferrer");
